fix/double r for big radius

+++ assignment_1/gauss-circle/is-gauss.h
@@ -1,3 +1,3 @@
 #pragma once
 
-int num_points(int r);
+int num_points(double r);
